FedGCL/FedGCL-main/requirements.txt
dgl_cuda11.0==0.9.0
numpy==1.21.6
pandas==1.3.5
scikit_learn==1.0.2
torch==1.10.1+cu111
torch_geometric==2.0.0
